tests/auth/requirements.txt
pytest==7.1.2
pydantic==1.9.0
python-dotenv==0.20.0
requests==2.28.1